chore: Move runner logic out of component
