+++ tests/mocks/runner.ts
@@ -1,0 +1,4 @@
+export class RunnerMock {
+  public run = jest.fn();
+  public stop = jest.fn();
+}
